cassie: manual tests and fixes against production data
Here is a summary of the main changes I had to do after Copilot's work.

Over 27 functions, 17 were dead-simple and worked or almost worked out
of the box. However for the following 10 functions, way more work was
needed to make them work:
* get_crashes_for_bucket.py: that one was just plain wrong and was
  rewritten
* get_metadata_for_buckets.py: without being wrong, any human would have
  seen that it just needed to call the previous function in a loop
* get_package_crash_rate.py: without being completely wrong, this one
  had to be fixed, because the AI doesn't know about `.order_by()`
* get_package_new_buckets.py: that one did work, but had Copilot really
  confused about str vs bytes
* get_retracer_counts.py: this one had half of it looking really weird,
  but that was dead-code anyway, so it's gone
* get_retracer_means.py: this one was rewritten from scratch. Copilot
  guessed whatever it wanted, but the new API just doesn't work like the
  old one, and I had to use a completely different approach.
* get_signatures_for_bug.py: copilot confused again with typing of int
  vs str
* get_user_crashes.py: some weirdness around cassandra and timestamp,
  that obviously got Copilot confused
* get_versions_for_bucket.py: Copilot used the wrong columns and
  reported a valid but completely different result from what was expected
* record_bug_for_bucket.py: usual mess with the str vs bytes vs int typing

In overall, I'm not sure regarding time gained or lost with regard to
using Copilot. What's sure is that I've been able to directly focus on
the interesting topic, without bothering with the initial boilerplate
of setting up harnesses to call each function, and the initial API
migration. I'd say it was a pretty positive experience, despite maybe
not looking like so from the PR or commit log.